Implement a button on the board.
We deleted the file and used a button to start a 10 second delay and activate the relay which activates the door strike.

When we program the relay for 1 second, it averages 7 seconds of total "on" time. When we tell it to turn on for 10 seconds, it stays on for 10 seconds exactly like it's programmed to.

Do relays have capactitors in them? We're confused.